+++ src/models/client_organization_domain_verification.rs
@@ -18,7 +18,7 @@ pub struct ClientOrganizationDomainVerification {
     #[serde(rename = "strategy")]
     pub strategy: String,
     #[serde(rename = "attempts", deserialize_with = "Option::deserialize")]
-    pub attempts: Option<i32>,
+    pub attempts: Option<i64>,
     #[serde(rename = "expire_at", deserialize_with = "Option::deserialize")]
     pub expire_at: Option<i64>,
 }
@@ -27,7 +27,7 @@ impl ClientOrganizationDomainVerification {
     pub fn new(
         status: String,
         strategy: String,
-        attempts: Option<i32>,
+        attempts: Option<i64>,
         expire_at: Option<i64>,
     ) -> ClientOrganizationDomainVerification {
         ClientOrganizationDomainVerification {
