+++ src/models/stubs_verification_saml.rs
@@ -34,14 +34,14 @@ pub struct StubsVerificationSaml {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub expire_at: Option<Option<i32>>,
+    pub expire_at: Option<Option<i64>>,
     #[serde(
         rename = "attempts",
         default,
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i32>>,
+    pub attempts: Option<Option<i64>>,
 }
 
 impl StubsVerificationSaml {
