+++ src/models/verification_google_one_tap.rs
@@ -20,9 +20,9 @@ pub struct VerificationGoogleOneTap {
     #[serde(rename = "strategy")]
     pub strategy: Strategy,
     #[serde(rename = "expire_at", deserialize_with = "Option::deserialize")]
-    pub expire_at: Option<i32>,
+    pub expire_at: Option<i64>,
     #[serde(rename = "attempts", deserialize_with = "Option::deserialize")]
-    pub attempts: Option<i32>,
+    pub attempts: Option<i64>,
     #[serde(
         rename = "verified_at_client",
         default,
@@ -43,8 +43,8 @@ impl VerificationGoogleOneTap {
     pub fn new(
         status: Status,
         strategy: Strategy,
-        expire_at: Option<i32>,
-        attempts: Option<i32>,
+        expire_at: Option<i64>,
+        attempts: Option<i64>,
     ) -> VerificationGoogleOneTap {
         VerificationGoogleOneTap {
             object: None,
